Add Bullet Behavior
The Bullet behavior simply moves parent object forwards at an angle. However, it provides extra options like gravity and angle in degrees that allow it to also be used. Like the name suggests it is ideal for projectiles like bullets, but it is also useful for automatically controlling other types of objects like enemies which move forwards continuously.