Add Snap to ground on Platform Behavior.
While the character is standing on the floor, it will try to remain snapped to it. If no longer on ground, it will not try to snap again until it touches down. The snap property (export) is a simple vector pointing towards a direction and length (how long it should try to search for the ground to snap).